ref: moving common widgets in widgets folder
